tagmate/requirements.txt
aiosqlite
arq
asyncpg
bcrypt
datasets
fastapi
fastapi-login
gunicorn
httpx
minio
pandas
passlib
python-multipart
redis
redislite
setfit
tortoise-orm
uvicorn
watchfiles



